Predictfluence/ds/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==2.0.36
psycopg2==2.9.10
pandas==2.2.3
scikit-learn==1.4.2
python-dotenv==1.0.1
